Add to the SimpleITK Java wrapping system.
The two things that need to be done in order to use the wrapped SimpleITK classes
from Java are to add the build directory to the class path and to load the
compiled libSimpleITKJava.so.  At this point I can't find a good automatic way
to do the former.  The classpath can be modified either by setting the CLASSPATH
environment variable or by using the -classpath "..." command line argument.

The library can be loaded using System.loadLibrary or System.load.  I've set up a
simple class in Wrapping/LoadSimpleITK.java.in that CMake will update to have the
proper build path.  In order to use SimpleITK, the user must simply call

LoadSimpleITK.load();

in their code.  I have also added a simple example in Examples/SimpleGaussian.java