Finished simpleAddImage and associated changes
While adding the simpleAddImage filter, and particularly the test for it, I ran
into a fairly major hurdle that needed addressing. Specifically, the current
simpleImage offered no method for accessing the data. To fix this, I made some
changes to the implementation of simpleImage and added a few convenient macros.

The ultimate goal was to give the user access to the pixels without requiring
knowledge of the pixel type. To this end I switched the constructor system so
that the only public constructors for simpleImage take itk::Image<type,3> as
the first argument rather than itk::ImageBase<3>.  The internal storage of the
image is still done as an itk::ImageBase<3>, but since only full Images can be
passed to the constructor, the data can safely be dynamically cast back to
its original type which can be retrieved directly from the image.

To complement this system I added void getPixel(int x, int y, int z, type* out)
methods for each supported pixel type this is slightly clunky and could also
be done using a single templated method of the form
"template<class type> type getPixel(...)" Either would work, but the former
continues to hide templating from the users.  To further cover this up, I added
several pixel access macros that give the user access to all pixels in sequence
with R, RW, and W access (called sitkForEachPixel<R/RW/W>).

Both the type constructors and the pixel access methods are populated using
macros. The only remaining issue seems to be a segfault that is generated when
using the java wrapping.