Initial simple registration class framework.
The Registration framework in SimpleITK consists of Registration, Transform and Resample classes.  Registration classes provide different transform/metric/optimizer/interpolation combinations.  This is a compromise on ITK's flexiblility, and could be explored and expanded in the future.  The only class currently offered is an Affine transform, Mutual information registration.  The output of this registration is handed off to the Transform class which stores a ITK base class.  The Resampler class has several resampling functions available, and more could be added as nessary.