Add initial support for C# wrapping
The bulk of this changeset is the addition of Wrapping/CSharpModules.
These CMake modules were based on those in the GDCM project.
So far only C# .NET on Windows is supported.
Still to be added is support for Mono on Windows and Linux.
Further work is required in SimpleITK.i for C# specifics (eg. rename toString).

Fix Visual Studio issues to do with stdint.h