Fixed bug in hash calculation, updated all tests with new hashes
In the MD5 and SHA1 calculations, the number of voxels in the buffer
was being used, however this should have been size of the buffer in
bytes.  Changed to use the number of voxels * sizeof the voxel data
type.  Also fixed bug where ValueType was being used instead of the
straight VoxelType.  We want the hash value to be computed on the
values in memory to avoid round off error, etc.  This will also avoid
problems cross-platform.  Hashes must be computed on the little endian
binary data.