Reworked how the executables are found now uses argv[0] and KWSys
Using SystemTools to tease apart the directory of the argv[0] was
sucessful.  When ctest runs from the IDE, it uses the current build
settings, from the command line, it searches.  Since ctest does all
the work, and we build all the executables and libraries in the same
directory, this works.