ENH: major revision to pixel typelists
Redid the pixel type list to support Scalar pixels, complex numbers,
vector images and label map images.

Added a classes to define pixel types for different image types. So
now the types in the list are not directly pixel types. This is  the
begining of convert from pixel type to PixelIDTypes.