ENH: Can set TestCase name in ImageCompare to re-use baseline images
Previously, we required a Python, Java, Lua, etc baseline for
ImageCompare.  Now, if the TestCase argument is not empty, use it to
construct the name of the baseline image.  This helps for generated
test code, we do not need to have separate baseline images for each
generate test in a wrapped langauge.