ENH: Add OperationType enum & support to MathematicalMorphology
Added support for a "public_declarations" field in the json.  This is
inserted verbatum into the generated header.  Using this mechanism, an
enum type is added to MathematicalMorphology and used as a parameter,
rather than unsigned int.

To support testing of this, the "tests" list is extended with a new
tag in the "settings".  Formerly the tag was a simple "parameter",
"value" tuple.  To enable custom types, support was necessary for
Python, Lua and Java.  Now during test generation, the Python template
looks for "python_value" in the settings list, falling back to
"value", if "python_value" does not exist.  Similar code is needed for
Lua, Tcl, Java, etc, because they have a language-specific
representation of enum types.  An example is below:

  "tests" : [
   {
     "tag" : "BinaryErosion",
     "description" : "Test binary erosion",
     "inputA" : "Input/STAPLE1.png",
     "settings" : [ { "parameter" : "Operation",
                      "value" :
   "itk::simple::MathematicalMorphologyImageFilter::BinaryErode",
                      "lua_value" :
   "SimpleITK.MathematicalMorphologyImageFilter_BinaryErode",
                      "python_value" :
   "SimpleITK.MathematicalMorphologyImageFilter.BinaryErode"
                    } ],
     "md5hash" : "9e37516c795d7f25847851666ef53ef9"
   } ],