ENH: Using baseline images for filters that are sensitive to round off
Different compiliers and machines will report different values for
these filters, so hash values are not useful.  Created baseline images
and tolerances for the filters.

Added testing for Index, fixed bug that dimensionality was reported as
2 in constructor for itk::Index<3>.

Nasty bug fixed where an image on disk is read with index [0,0,0], but
ExtractImageFilter would return an image with index [0,0,slice] where
slice was the extracted slice.  This caused SubtractImageFilter to
fail with an exception, but for some reason I could not get GDB to go
into itk::DataObject to debug.  Caught it the old-fashioned way,
printf.

Added testing of default values to the tests to increase coverage.