BUG: Work around [[]] conflict between R and Lua.
The Lua expand script uses the symbols [[ ]] and gets confused with
the code in R that uses the same symbols for arguments.

Rewrote the test in order to extract R arguments into temporary
variables and in that way avoid having [[ ]] symbols inside of the
code that the source code generator must expand.