FIX: Fixed java wrapping
The first fix needed was to change the location of libSimpleITKJava.so from the
Wrapping directory to the bin directory in LoadSimpleITK.java.in. I also
updated the name of the build target from org_simple_itk_jar to
org_itk_simple_jar. Finally, I also updated the SimpleGaussian.java example to
use the new method names (with the first letters capitalized).

NOTE: In order for the java wrapping, the install target must be built. This
will require whatever permissions are necessary to access the directory in
which the jvm is installed (in my case it needed admin permissions)