ENH: added check for tr1 functional
Added CMake check for tr/functional header. We avoid assuming a
specific compiler is configured a certain way, and instead utilize try
compiles with CMake to perform configurations.

Check for "functional" and "tr1/functional", prefering the later since
it qualified with a directory and may prevent conficts in some
situations.