ENH: Removed unneeded loading code (was broken anyway)
Instead of explictly loading a shared library, now reley on the Java
path to load the library.  Generally, this will examine the user's
"PATH" variable in a system-dependant fashion and try to load a shared
library.  For testing, we explicity set the java.library.path variable
to point to the binary directory of our build.