ENH: Support for Ruby wrapping
Ruby wrapping is still experimental at this point, and should be used
with caution.  There are some fairly complex requirements to build
Ruby modules, and it involves SWIG re-writing method names.

UseSWIG.cmake was modified to UseSWIGLocal.cmake to include support
for Ruby modules.  Still needs to be tested on Linux and Windows
platforms.  Testing was augmented with paths to Ruby executables.