ENH: Modified the template to erase previous images
If the test fails to write output, the driver can mistakenly mark the
test as passed.  This change makes the driver write over the image
with a small file that should fail the test if not overwritten by the
correct result.

Would be nice to erase the temp directory before running the tests,
but not sure how to do this.