FIX: Clean java build when Code changes
This isn't the cleanest solution, but we needed a way to delete all of the
generated .java and .class files when things changed in any of the Code
libraries. We also needed to force regeneration of SimpleITKJava_wrap.cxx. The
potential problem with this fix is that it relies on unix commands to make sure
that there are files to delete in Wrapping/org/itk/simple and
Wrapping/build/org/itk/simple which may cause platform dependency issues