ENH: Added PixelContainer interface.
The PixelContainer interface is the carrier of the ITK ImportImageContainer.
It is mainly intended to facilitate the connection between SimpleITK image
and existing applications that use their own representation for their image
class. This includes exporting the pixel buffer to data structures of languages
targeted through wrapping (Python, Java,...).

The PIMPL implementation of the sitkImage was reoganized in order to take
the implementation of templated instances out of the header file. The pixel
container interfaces follows now a simimlar PIMPL distribution.