ENH: Implemented simple CastImageFilter with dual member factory
This change enabled the casting of vector image to vector image,
basic pixel to complex numbers, along with the existing, basic pixels
to label maps and basic pixels to vector images.

The key feature used was the ability to specify the input and output
type lists for methods, along with an additional parameter to specify
the addressor for the member function.