COMP: fixing pimple implementation for VS
Deriving from a protected in class class definition is not allowed by
C++. Instead we use a forward declaration in the header, and declare
the private implementation in the cxx/txx file. This is standard for
the pimple idiom, but the templated class makes the instantiation more
challenging.

To instantiate the PixelContainer's templated pimple, they are
implicitly instantiated via the sitkImage's pimple classes. This makes
for a very lean sitkImage.h and sitkPixelContainer.h files.