COMP: Fix wrapped source dependencies
The SimpleITKXXX_wrap.cxx files all now depend properly on the libraries and
will be rebuilt whenever one of the libraries has changed. The only remaining
custom cleaning command is for the java source and class files. I wasn't able
to find an alternate solution to this that did not run every time make was
called