ADD: Intial registration metric class (Mattes)
The Mattes MI registration metric is the first prototype class for
registration.  It operates by creating an
itk::MattesMutualInformationImageToImageMetric on demand, based on the
input image types.  The result is passed back as the non-templated
super class, itk::SingleValuedCostFunction.  This allows a different
class to construct a registration object from generic components.