ENH: Added super class hierarchy and wrapping support
Added Interpolate, Metric, Optimizer, and Transform super classes to
the hierarchy.  The mechanism for assembling a registration object is
to create objects of the required type for the 4 components.  For
instance, AffineTransformation, RegularStepGradientDescentOptimizer,
LinearInterpolate, MattesMutualInformationMetric.  Set these objects
in the Registration object, and run the registration with the execute
method.

Output of registration is a std::vector<double>.  These are the
parameters to a Transform of the type given to the Registration
object.

This commit includes a test.