ENH: Detecting 64 bit compiles to aid SWIG wrapping
SWIG generates incorrect code on 64 bit compiles when using int64_t,
because it uses a typedef of long long, which works in 32-bit
compiles.  Using the size of a void* to indicate 32- or 64-bit
compile.

Also fixes the dependancies on language-specific .i files.

SIMPLEITK-34 #comment Need to use 32- and 64-bit specific define.
SIMPLEITK-15 #resolve Added dependancies for .i files