ENH: improved C++ gaussian example
Changed from the one directional RecursiveGaussian filter to the
SmoothingRecursiveGaussian. This changed the result of the programs.

Addiionally, added a cast from pixel type of the output of the
smoothing filter to the type of the input. This help enable writing
back to the same file type as the input.

Added more comments too.