ENH: Small refactor of registration, fixed Optimizer namespace bug
Refactored registration framework to return a pointer to a transform.
The transform is the same used in the registration, but has parameters
set properly.  Fleshed out Transform base class.

Had a nasty bug in the optimizer hierarchy.  Because sitkOptimizer.h
was not included in sitkRegularStepGradientDescentOptimizer.h, it was
actually deriving from itk::Optimizer, which caused all sorts of
problems!  This is now fixed, and working properly.