COMP: 57 - Properly reset GeneratedSource list in cache
The SimpleITKBasicFiltersGeneratedSource list was not getting cleared out in
the cache between configurations, so it was continuously being added to. This
was directly causing a problem when switch what filters were being generated on
different branches, but it was also a problem because the cache was growing
out of control even with out fixing branches. All fixed now

SIMPLEITK-57