ENH: broke the instantiation of the CastImageFilter in 6 files
Added new private methods
CastImageFilter::RegiterMemberFactor[23][vl]?() to break the
instiation into multiple files. Each function is in a separate
compilation unit. It is split along dimensions and Image type.

This is expected to speed up compilation times on multi-core
systems. Additionally, distcc was timing out due to the long
compilation of the single file.