ENH: Added support for vectors of bool
Added wrapping support for std::vector<bool> with the std_vector.i
swig modual.

Modified typed languages to support filter members with type
vector<bool>.

Note: SWIG did not appear to create a VectorBool::push_back method
for JAVA, this is unclear as to why that is. The JAVA test template
was changed to pre-allocate and use the set methods.

SIMPLEITK-63