ENH: added try compile to test if uint64_t and unsigned long are same
Swig needs SWIGWORDSIZE64 defined when the uint64_t is of type long,
otherwise it uses type long long. Please remember that even if these
types are the same length they are different intrinsic types, and the
templates are not equivalent. A CMake try compile is used to make this
check at compile time and the flag should be correctly defined.

SIMPLEITK-37