COMP: 86 - Fix tests not building and wrapped libraries not found
When "subdirs" was switched to "add_subdirectories" the order in which Testing
got visited changed, so it was being visited before the json files had been
parsed.

Also, the libraries were moved to the "lib" directory instead of the "bin"
directory. This change was not reflected in the library linking for the wrapped
tests.

SIMPLEITK-86