BUG: test for functional unordered_map
Added try compile for functional stl tr1 unordered_map. On Gcc 4.0 on
mac OSX, there is an internal constant correctness error for certain
function calls. A try compile was added for this.

Addtionallally a configure defined " SITK_HAS_UNORDERED_MAP" was added,
to simplify the logic of when this type can be used.