ENH: enabled multi-input filters to work with different size images
While most binary image filters don't work with images of different
size, some to. The test was changed to enable some filters to pass the
test when they can work with images of different sizes. The test now
just tries to run, and either an excetion or running is OK. So this
part of the test only check to see that the behavior is defined.