ENH:  Add infrastructure for Numpy conversion.
A custom method is added to the SimpleITK C module for converting Image's to and
from numpy arrays.  Placeholders are added for now.  The code must be included
inline because the function are declared static.

CMake is then used to append the methods to the SimpleITK.py module.

A test is started.