ENH: Implement GetArrayFromImage.
Implementation with a hybrid of C Python and interpreted Python.  The
implementation should build without numpy necessarily available, but fail at
runtime.  The use of the Py_buffer required Python 2.6 or better.  Some pixel
types are not supported because of todo implentation in the sitk::Image code.

The test sometimes segfaults.