ENH: Disable Launchers usage
Applying patch recommented,

On Sep 2, 2011, at 11:56 AM, David Cole wrote:

One thing we've noticed looking at scripts in other projects is:

- one project that also uses a SuperBuild technique uses
CTEST_USE_LAUNCHERS=1 but they also have a subproject-organized
dashboard

- another project using a SuperBuild has this in its dashboard scripts:

 # Warning - enabling launchers causes linux not to submit warnings
or errors to cdash
 set(CTEST_USE_LAUNCHERS 0)

For now, I'm going to recommend that we hard-code CTEST_USE_LAUNCHERS
to 0 for the simpleitk_common.cmake script. Since I do not have push
privileges to the 'dashboard' branch, I'll ask you to make that
change. Rather than allowing it to be set from outside by the
containing script, just always set it to 0 for SimpleITK
superbuild....

This should cause warnings to show up after the machines update to
that version of the script.

Thanks,
David Cole