ENH: Finish implementation of GetImageFromArray.
Implement all supported pixel types and dimensions.

Fix the shape/size spectifications so they are correct for numpy/sitk::Image
spectifcations.  Note the numpy shape is like the order when specifying the size
of a C array while the itk::Image is displayed in reverse, i.e. [x, y, z].

Tests are implemented and updated.