ENH: Ripped out dependencies on ITK headers in header files
Minor changes were need in several places. Use stdinint.h instead of
itkIntType.h and using a forward declaration of Smartpointers were
needed.

The largest change was ripping the superclass itk::Execption for the
sitk::Exception class. The sitk exception is now directly derived from
the std exception class, currently it is void of content and need to
be reimplemented.