Flag SWIG to generate Python documentation with argument types
SWIG does not, by default, generate docstrings for Python that contain
the object types.  Adding the "-features autodoc=1" enables this for
all Python functions and classes.  For example:

Help on function Abs in module SimpleITK:

Abs(*args)
    Abs(Image arg0) -> Image

    Computes the ABS(x) pixel-wise.

    li {ImageProcessing/AbsImageFilter,Compute the absolute value of
    an image}

This gives a clue to the user how Abs should be used, i.e., that it
takes only one Image argument.