Fix SimpleITK C# x64 platform issue.
At the moment when a C# assembly is compiled on an x64 platform a CS1607
warning is displayed (referenced assembly targets wrong platform).
This is because for x64 platforms the wrong compiler is found
(the {windir}\Framework\..\csc.exe is used, rather than
{windir}\Framework64\..\csc.exe).

It seems difficult and inelegant to start detecting different compilers
based on platform, so the solution employed here is to always compile
C# assemblies for ANYCPU (this is the current solution for Linux systems
anyhow). This means the the interpreter will automatically generate the
correct platform-specific machine code (x86 or x64) on-the-fly.

In addition to this fix, some minor refactoring is done including:
(a) adding "call" to the front of Windows Mono compiler is now done in
FindCSharp rather than UseCSharp,
(b) minor refactoring for "if" statements to make more readable,
(c) append references/sources using strings rather than lists, and
(d) added status messages to see the compiler command line (useful
for debugging).

Change-Id: I8b3df041852347113b6da6c162f73eb44152a86c