ENH: add new constructor to Image which takes vector and numcomponets
Previously there was not constructor where you could specify the
number of components for a vector image. There would be ambiguilty if
it was just added as another scalar parameters. So a new method was
added where the size is a vector and the last parameter is an optional
number of components.