Fix compilation of C# examples.
The C# example (SimpleGaussian) was not being compiled.
Even though add_custom_command and add_custom_target were being called,
the command was not added to Visual Studio projects by CMake.
The issue seems to be that the add_custom_command OUTPUT variable needs
to be an absolute path. The add_custom_target DEPENDS variable is also
changed to match.

In addition, the name of the C# example is changed to
SimpleITKCSharpExampleSimpleGaussian to make clear the purpose of the
custom target.

Finally, a previous commit fixed the use of "safe" C# compiler path
for Win32 systems (by adding "call" before the compiler executable),
so this is removed from Testing/Unit/CMakeLists.txt.

Change-Id: I771609a360d242971f64b78fec0b21db2cb3839a