ENH: improved ctest test name for new sub-gtest
The name was the same as the gtest search expression. This was ugly,
so not the first part of the expression is used as the name of the
ctest.

Note: There is some overlap with some of the tests and these glob
expressions. This is with the *_ENUMCHECK manually written test.