ENH: Added CMake macro for testing CSharp examples and test
A manually written C file needed to be written for each custom CSharp
example or test, and additionally CMake code was needed to compile
it. This has been simplified to just need to use a cmake macro which
compiles and adds the test. Additionally, this uses the itkTestDriver
which support a variety of image comparision operations.