BUG: CSharp variable need to propagate from super build to simpleitk
There were not CSharp cmake variable marked to be propagated from the
Superbuild to the SimpleITK external project. Specifically the newly
added CSHARP_PLATFORM needed to be set and it was not work. However,
in the FindCSharp.cmake file, many of these variables are unset. This
may unset the propagation which occours though cmake -C and -D command
line options.