Correct R_values in json file.
With this patch and the RSugar topic, most tests are passing. This
correction of values should stand alone and improve the current test
even with out the R Sugar.

Unit testing.
============

Good news:

97% tests passed, 15 tests failed out of 430

Without the swig changes you should get a couple more failures. I don't
know my way around the testing architecture well enough yet to figure out
how to fix all of the problems, but changing some of the test files manually
does fix the errors. I've also added a hack to RImageFilterTestTemplate.R.in
that fixes the problems with true/false instead of TRUE/FALSE, until that
can be corrected in a more appropriate place.

Failures are:

	296 - R.Extract (Failed) -- collapse strategy should be a
                                    string 'DIRECTIONCOLLAPSETOIDENTITY'
                                    Fixing manually leads to test passing.
	277 - R.ConnectedThreshold (Failed) -- AddSeed(SimpleITK.Index{}) -
                                               should be c(100,100)
	300 - R.Flip (Failed)
	299 - R.FastMarching (Failed) -- setting seeds - some sort of table
	250 - R.BinaryDilate (Failed) -- looks OK to me - possible test fault
	380 - R.RegionOfInterest (Failed) -- SetSize
	379 - R.RecursiveGaussian (Failed) -- enumeration
	334 - R.LabelMapToRGB (Failed) -- test fault Pixel type: 8-bit unsigned
                                       integer is not supported in 2D
                                       byN3itk6simple24LabelMapToRGBImageFilterE
	333 - R.LabelMapToLabel (Failed)
	332 - R.LabelMapToBinary (Failed)
	331 - R.LabelMapOverlay (Failed)
	330 - R.LabelMapMask (Failed)
	327 - R.IsolatedWatershed (Failed) -- filter$SetSeed2([ 120, 99 ] )
	373 - R.Paste (Failed) -- filter$SetSourceSize([ 20, 15 ] )
	365 - R.NeighborhoodConnected (Failed) -- seeds

I've gone through the json files and reduced the failures to the following:

	300 - R.Flip (Failed) -- swig problem -- need vectors of bool
	299 - R.FastMarching (Failed) -- needs vectors of vectors
	334 - R.LabelMapToRGB (Failed)
	333 - R.LabelMapToLabel (Failed)
	332 - R.LabelMapToBinary (Failed)
	331 - R.LabelMapOverlay (Failed)
	330 - R.LabelMapMask (Failed)

Change-Id: Id58efc24a9f202232100edcc35885511bc40ee10