Enabling R testing by fixing loading, and correcting template
Added a script which is configured by cmake, to manually load the
simpleITK shared library and the R file. This is then loaded by the
test, by a set enviroment variable.

The Rscript command was not functioning correctly on my system. The
tests are now driven with R proper executable with arguments to run it
in a scripted fashion.

Corrected the R test template to actually work. Much of the code
appears to have been copied from Python.

Change-Id: I83db1c21494978f1a42f5067753a0ad9d28ddc58