Coverted system call to kwsys Processing functions.
The "system" call is not portable because it relies on the system
shell, and issues related to handling of spaces was problematic on
windows. This approach uses kwsys's Process functiontion to create a
new process, and check to see if there is an initial error, the it is
detached from the calling.

Change-Id: I5c096b0f01bf4eff36665c62d1049ebc1fd07657