Enabled image compare methods to work with vector images
A special case was made to calculate the RMS of vector images in the
tolerance image compare method. This calculated as the average of the
difference if the components then the RMS is computed.

Additionally, the scalar images are now cased to float before
squaring, it's possible that there was previously numeric overflow
causing errors before.