Broke BasicFilters library into a set of libraries
This library in now automatically broken up into libraries so that
there are only N compiled files in each library. The libraries are
aded to the now cached interal varible for the list of SimpleITK
libraries. Additionally, the ITK libraries were removed from this
list, since if needed they will be transitivly linked.