only add warning flags to SimpleITK build
Previously the warning flags were passed to ITK, the prior change to
Swig and PCRE also passed all the warning flags to them aswell. This
causes too many warnings, so the additional warning flags are only
passed to SimpleITK.

Additionally, the tested compiler flags settings that are for warnings
are explicitly set to CXX_ADDITIONAL_WARNINGS_FLAGS, to clearly
indicate the purpose of those additional flags.