Refactored Show for readability & added color support.
Code from Show was pulled into separate functions BuildFullFileName and FindApplication to
untangle the code.

Also Show will now display color images in ImageJ.  If we detect a 3 channel,
unsigned 8-bit or 16-bit image, it is automatically displayed in color.  We
do this by invoking ImageJ with a macro that says to combine the channels
into a composite image.

Cleaned up #ifdefs

removed use of __linux__

Change-Id: I209c47022e6811aa5a44d0852766d3a6a7d7452f