Put manually created Fitlers in separate library
Some of the expand template generated filters depend on files that are
manually generated and they need to be linked against. So, all the manually
created basic filters files are placed in BasicFilters0, and the other
libraries are linked against it. The one limitation with this approach is
that there can be no inter-dependencies, or every inter includes with the
generated files in the basic filters.