Refactoring to simplify boundary conditions support
Moved the switch statement for selecting the boundary condition
to a function in a source file callable from the custom_itk_cast
attribute for the BoundaryCondition member. Modified the json
files accordingly.

Added boundary condition choice to convolution filters

The convolution filters now follow the model of the deconvolution
filters in providing a choice of boundary conditions.

Change-Id: I7566db649a6293342f82cbac46389f6a7ad92813