BUG: The BasicFilter libraries were not being added to cmake configs
After the list of libraries became dynamically generated, it was not
correct in the Cmake configure file for other to use the project. The
time when the config file was configure was before the cmake variable
was set in the sub-directories. It is now done at the end of the cmake
file and is correct.