ENH: Improved R packaging
This commit includes a directory heirarchy that is an R package
template. It doesn't quite conform to the R standards, as it
is designed to create an installable R binary package, rather
than being a source package. The latter will be necessary as
we look at getting the bindings onto CRAN.

Additional commands in Wrapping/CMakeLists.txt copy the
package heirarcy to SimpleITK-build/Wrapping/Rpackaging, then
copy the swig generated R file and the shared library into
the appropriate subdirectories. The package is the installed
into SimpleITK-build/Wrapping/RLib and built into a tarball
in SimpleITK-build/Wrapping/. The former will be used when
running tests, the latter can be distributed or installed.

I have also tried to eliminate the search for libR.so/libR.dylib
etc from FindR.cmake and SITKLanguageOptions.cmake, as the shared
library for R isn't always installed and isn't needed for package
building.

The vignette building process isn't clean yet. I don't understand
why it doesn't work. R allows a pdf to be included, and that's
what I have done.

Change-Id: I148e3bd6672a8ae5eb74373d1cd9421d322612ac