BUG: FastMarching test in R
This test requires that a set of TrialPoints be set. Other languages
have a special case and generate a series of calls to AddTrialPoint.
This is obviously feasible in R (as the same approach is used for
SeedList). However it doesn't test the binding between R and methods
that can take vectors of vectors, which R can now handle of lists
of vectors. One option is to construct the list via use of another
special case like this:

elseif parameter == "TrialPoints" and R_value then
	    OUT=[[
	          filter <- filter$$ClearTrialPoints()
	          tp <- list()
	          $(for  i=1,#value do
	            OUT=OUT .. "tp\[\[" .. i .. "\]\] <- " .. R_value[i] .. " \n"
	            end)
	          filter <- filter$$SetTrialPoints( tp )
	          ]]

But this leads to a special branch for each new method with a similar pattern.

The alternative option implemented in this commit is to construct the
R list as a string in the json file. This wasn't quite right before as the
square brackets were still present. This approach requires some R knowledge
when setting up the tests in json files (which is required anyway), and
should be independent of method name. The tests with SeedPoints may
be able to be modified in similar ways, leading to a simpler test template.

Change-Id: I8ec627ced8bc850f505e0ce617fafde4635ef590