ENH: Basic arithmetic/maths overloading for R
R operators, such as comparisons, logic, arithmetic and maths functions
can be overloaded for arbitary classes. This patch allows the standard
maths operators to call the appropriate underlying simpleITK functions,
much as can be done with the python binding. e.g.

a <- ReadImage('cthead1.png')

b <- a + 20
g <- a + a
h <- a - a

k <- sqrt(a)
k <- sin(a)

etc

Added ! and xor.

Change-Id: I9b0f440b9da5587ada223678188d13883bab2ff5