Compare what filters ITK and SimpleITK implement.
CompareITKandSITKFilters.py is a python script that compares what image filters
ITK (via WrapITK) and SimpleITK implement in python.

The script works by getting all the entries from the python modules and looking
for the ones ending in "ImageFilter" (and "ImageSource" for WrapITK).

The results are written to a CSV file.  The default name is "filters.csv".
This file can also be used as input to the script.

Added a ToDo boolean and CSVtoTable.py

Added a ToDo boolean flag for filters in ITK but not SimpleITK.

Created CSVtoTable.py, a script that takes an output file from
CompterITKandSimpleITKFilters.py and produces a HTML table.  By
default CSVtoTable.py reads in the file "filters.csv" and produces
the file "filters.html".

Checked in the filters.csv file.

Change-Id: I2f3536bd880d8adacf18d2075f0393c1adaf59a9