Removed the "%m" macro token.
For displaying color images with ImageJ, I create a macro that tells ImageJ
to combine the components of a vector image into the channels of a color one.

Previously this macro could be accessed by the user in one of the command
environment variables, such as SITK_SHOW_COLOR_COMMAND, via the "%m" token.

This access is not needed and makes the code more confusing.  So I have
removed it.  Now the default values for commands that need a color macro have
the macro explicitly stated.  For example, SITK_SHOW_COLOR_COMMAND on
a Linux build is:

    static std::string ShowColorImageCommand = "%a -e "
        "\'open(\"%f\"); run(\"Make Composite\", \"display=Composite\"); \'";

Previously the second part of the string was simply "%m".

Change-Id: I18bc5d9ace701973324abdaf5b87a31a87c3c7af