ENH: R - import vector image buffers.
The C code has been modified to allow import of vector buffers. The
R method to use this enhancement looks like this, and will be
included once the R packaging is organised.

setMethod('as.array', "_p_itk__simple__Image",
          function(x, drop=TRUE) {
            sz <- c(x$GetNumberOfComponentsPerPixel(), x$GetSize())
            ans = .Call("R_swig_ImAsArray", x, FALSE, PACKAGE = "SimpleITK")
            dim(ans) <- sz
            if (drop)
              return(drop(ans))
            return(ans)

            }
          )

The only changes to the C side are including the number of components
in calculation of the buffer size, and including the vector types
in the switch statement.

Change-Id: I7a9fbf3b924740baa7150cbcf9c83db0416ce1de