Added Doxygen support.
CSVtoTable.py now has a "-d" option to output a Doxygen header and tail.

Doxygen.cmake was modified to automatically run CSVtoTable.py when
building documentation.  "filters.csv" in the source tree is taken
as input, and the output is written to FilterCoverage.dox in
the build tree under SimpleITK-build/Documentation/Doxygen.

Change-Id: If089874d40be5a57c0c7c4b5a6e3303ba7d1858d