Significantly increase the tolerance for TernaryMagnitude tests
The tolerance in not normalizes so about 10% of the pixel are off by a
floating point bit, which maps to 256.0. This is due to the nature of
squaring data that covers the short range, and squaring the
difference. Really, this is not a very stable operation. to the
tolerance needs to be extended to cover the reported 94 RMS.