Fix runtime error in extracting slice.
When a test fails the Normalize and Save method may try to save an
slice for display on the dashboard. This was failing with 3D images
because the collapsing slice size was not set to zero.

Also added exception handling where the NormalizeAndSave methods was
used to preven unhandled exceptions.