Adding a variable to name SWIG library targets
Previously the name of the library targets may collide, which caused
interesting errors during build from multiple swig libraries having
the same name.

Now the swig language is added to the target name, and a new CMake
variable SWIG_MODILE${name}_TARGET_NAME is introduced so that
additional properties can be added to the target. However, there is
still a name collision with this variable, but that does not appear to
be an issue yet.