Change python _GetImageFromArray to _SetImageFromArray
More code was extracted from the C python code and placed as python
code. So the C method now only fills the sitk Image's buffer with the
provided array. Priviously the array was provided by the python buffer
interface ( PEP 3118 ), this interface was not provided by numpy in
OSX 10.6 nor in Slicer, so with less effort the older read buffer
interface use used via ParseTuple arguments.