Properly implemented title for Show
Previously the title argument for the Show method wasn't used in the correct
manner.  It was used to create the temporary file name which ImageJ then
happened to use as the title of its window.

Now the title is correctly passed to ImageJ via the rename function.

Also previously if the title string had any spaces, ImageJ was not
launched properly.  Now the temporary file name created from the title
string has whitespace removed from it.

And also now one of the Show invocations in sitkShowTest.cxx includes
a title string with whitespace in it to test these changes.

Change-Id: I3498fd8325716c8250b7f0a92d49788426454435