ENH: Add support for C# SDK version.
Previously, by default, the CMake scripts would detect available
C# versions and default to the latest version (typically v4.0).
On a Windows system, v4.0 can only be used by Visual Studio 2010
(which is not desirable).

This changeset addresses the issue by adding support for the C#
SDK version. Unfortunately .NET and Mono handle this differently.
For .NET we must use the csc.exe from the desired framework
directory (done in FindDotNetFrameworkSdk.cmake). For Mono we
can simply use the commandline switch /sdk.

With this changeset the CMake scripts default to SDK v2.0.
This means that now the precompiled binaries can be used with
Visual Studio 2008 and later.

NOTE: Some of the changes may conflict with commits already on
next branch.

Change-Id: Ic7b72b947e070672b41811419a0c9ab5fb820fcf