COMP: Add special config for building SimpleITK w/Clang++
CLang++ requires a disambiguation of template elaborations by use
of the template keyword.

See this bug report for a description of the problem:
https://issues.itk.org/jira/browse/SIMPLEITK-296

Since Visual Studio can't handle addition of the template
keyword in this context, and GCC doesn't apparently need it,
I added a preprocessor macro so that the keyword is only used
if the compiler is CLang.