ENH: improving python egg version to use PEP386
The version should be consistent and sequential along the release and
master branches. A comparison is made between the version recoreded in
CMake and the version of the most recent tag. If there are the same
then the "post" suffix is used other wise the "dev" suffix is
used. dev meanin that we are working towards, where as "post" we are
patching away from.

In the future it would be good to detect which branch is the HEAD is
currently at and special case this.